A new image: canonicalwebteam/node
This is forked from "canonical-webteam/yarn", but removed the
functionality to run `yarn` by default.

As this image now includes support for bower as well, instead of
considering it a "yarn" application image, I think we should consider it
a "node" one. It's for encapsulating node modules, including "yarn" and
"bower" which will be preinstalled.

This change in structure means that we should no longer make "yarn" the
default command to run, and instead if you want to run yarn, you should
run it explicitly (`docker run canonicalwebtea/node yarn`)